+++ base/src/main/java/io/github/sds100/keymapper/base/actions/ActionOptionsBottomSheet.kt
@@ -24,14 +24,12 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.SheetState
-import androidx.compose.material3.SheetValue.Expanded
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
@@ -413,8 +411,8 @@ private fun Preview() {
     KeyMapperTheme {
         val sheetState = SheetState(
             skipPartiallyExpanded = true,
-            density = LocalDensity.current,
-            initialValue = Expanded,
+            positionalThreshold = { 0f },
+            velocityThreshold = { 0f },
         )
 
         ActionOptionsBottomSheet(
@@ -472,8 +470,8 @@ private fun PreviewNoEditButton() {
     KeyMapperTheme {
         val sheetState = SheetState(
             skipPartiallyExpanded = true,
-            density = LocalDensity.current,
-            initialValue = Expanded,
+            positionalThreshold = { 0f },
+            velocityThreshold = { 0f },
         )
 
         ActionOptionsBottomSheet(