+++ system/src/main/java/io/github/sds100/keymapper/system/permissions/Permission.kt
@@ -17,4 +17,5 @@ enum class Permission {
     ANSWER_PHONE_CALL,
     FIND_NEARBY_DEVICES,
     POST_NOTIFICATIONS,
+    READ_LOGS,
 }