+++ system/src/main/java/io/github/sds100/keymapper/system/settings/SettingType.kt
@@ -1,0 +1,12 @@
+package io.github.sds100.keymapper.system.settings
+
+import androidx.annotation.Keep
+import kotlinx.serialization.Serializable
+
+@Serializable
+@Keep
+enum class SettingType {
+    SYSTEM,
+    SECURE,
+    GLOBAL,
+}