+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## [4.0.0 Beta 3](https://github.com/sds100/KeyMapper/releases/tag/v4.0.0-beta.03)
+
+#### TO BE RELEASED
+
+## Added
+- #1871 action to modify any system settings
+- #1221 action to show a custom notification
+- #1491 action to toggle/enable/disable hotspot
+
+## [4.0.0 Beta 2](https://github.com/sds100/KeyMapper/releases/tag/v4.0.0-beta.02)
+
+#### 08 November 2025
+
+## Added
+
+- #1890 add button to save log to file and share it. The clipboard button now cuts off older entries and keeps newest ones.
+
+## Fixed
+
+- Only autostart PRO mode with Shizuku if Shizuku permission is granted. Otherwise fallback to method with Wireless Debugging and WRITE_SECURE_SETTINGS permission.
+- Starting system bridge for the first time would be janky because granting READ_LOGS kills the app process. Only grant for READ_LOGS when sharing logcat from settings.
+- #1886 mobile data actions work in PRO mode.
+
 ## [4.0.0 Beta 1](https://github.com/sds100/KeyMapper/releases/tag/v4.0.0-beta.01)
 
 #### 01 November 2025